+++ src/OrderedList.php
@@ -155,7 +155,7 @@ final class OrderedList implements MemberList
      *
      * @return array<int, float|int|bool|string>|float|int|bool|string|null
      */
-    public function value(string|int $offset): array|float|int|bool|string|null
+    private function value(string|int $offset): array|float|int|bool|string|null
     {
         try {
             $member = $this->get($offset);